requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
qrcode[pil]
python-jose[cryptography]
passlib
